Modals added

+++ client/src/app/account/account-nav/account-nav.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
+import { ConfirmService } from 'src/app/core/services/confirm.service';
 import { AccountService } from 'src/app/landing/account.service';
 
 @Component({
@@ -11,17 +12,18 @@ import { AccountService } from 'src/app/landing/account.service';
 export class AccountNavComponent {
 
   constructor(public accountService: AccountService, private router: Router,
-    private toastr: ToastrService) {}
+    private toastr: ToastrService, private confirmService: ConfirmService) {}
 
   onClickDeleteAccount() {
-    this.accountService.deleteAccount().subscribe({
-      next: () => {
-        this.toastr.success('The account was deleted and removed successfully');
-        this.accountService.logout();
-        this.router.navigateByUrl('/sign-in');
-      },
-      error: () => {
-        
+    this.confirmService.confirm().subscribe({
+      next: modal => {
+        modal && this.accountService.deleteAccount().subscribe({
+          next: () => {
+            this.toastr.success('The account was deleted and removed successfully');
+            this.accountService.logout();
+            this.router.navigateByUrl('/sign-in');
+          },
+        })
       }
     })
   }
